Make <SidepanelDataView> single-purpose
