+++ arches/app/src/components/ControlledListManager/SidepanelDataView.vue
@@ -1,57 +1,40 @@
 <script setup lang="ts">
 import arches from "arches";
+import Cookies from "js-cookie";
 import { computed, ref } from "vue";
 import { useGettext } from "vue3-gettext";
 
 import DataView from "primevue/dataview";
+import { useToast } from "primevue/usetoast";
 
 import SearchAddDelete from "@/components/ControlledListManager/SearchAddDelete.vue";
 
 import type { Ref } from "@/types/Ref";
-import type {
-    ControlledList,
-    Selectables,
-} from "@/types/ControlledListManager";
+import type { ControlledList } from "@/types/ControlledListManager";
 
 const lightGray = "#f4f4f4";
 const slateBlue = "#2d3c4b";
-const { $gettext } = useGettext();
+const ERROR = "error";
+const { $gettext, $ngettext } = useGettext();
+const toast = useToast();
 
-const props: {
-    addLabel: string;
-    createAction: () => Promise<void>;
-    countLabel: string;
-    delAction: (selected: Selectables) => Promise<void>;
-    delLabel: string;
-    delLabelPlural: string;
-    fetchAction: () => Promise<void>;
-    noSearchResultLabel: string;
-    noSelectionLabel: string;
-    selectables: ControlledList[];
-} = defineProps([
-    "addLabel",
-    "createAction",
-    "countLabel",
-    "delAction",
-    "delLabel",
-    "delLabelPlural",
-    "fetchAction",
-    "noSearchResultLabel",
-    "noSelectionLabel",
-    "selectables",
-]);
-
+const lists: Ref<ControlledList[]> = ref([]);
 const displayedList: Ref<ControlledList | null> = defineModel();
 const selected: Ref<ControlledList[]> = ref([]);
 const searchValue = ref("");
 
-const filteredSelectables = computed(() => {
+// Strings: $gettext() is a problem in templates given <SplitterPanel> rerendering
+// https://github.com/archesproject/arches/pull/10569/files#r1496212837
+const NO_MATCHING_ITEMS = $gettext("No matching items.");
+const NO_SELECTION_LABEL = $gettext("Click &quot;Add New Item&quot; to start.");
+
+const filteredLists = computed(() => {
     const loweredTerm = searchValue.value.toLowerCase();
     if (!loweredTerm) {
-        return props.selectables;
+        return lists.value;
     }
-    return props.selectables.filter((selectable) => {
-        return selectable.name
+    return lists.value.filter((list) => {
+        return list.name
             .toLowerCase()
             .includes(loweredTerm);
     });
@@ -66,7 +49,7 @@ const toggleCheckbox = (selectable: ControlledList) => {
     }
 };
 const selectAll = () => {
-    selected.value = props.selectables;
+    selected.value = lists.value;
 };
 const clearAll = () => {
     selected.value = [];
@@ -75,27 +58,116 @@ const selectRow = (selected: ControlledList) => {
     displayedList.value = selected;
 };
 
-await props.fetchAction();
+
+const fetchLists = async () => {
+    let errorText;
+    try {
+        const response = await fetch(arches.urls.controlled_lists);
+        if (!response.ok) {
+            errorText = response.statusText;
+            const body = await response.json();
+            errorText = body.message;
+            throw new Error();
+        } else {
+            await response.json().then((data) => {
+                lists.value = data.controlled_lists;
+            });
+        }
+    } catch {
+        toast.add({
+            severity: ERROR,
+            summary: errorText || $gettext("Unable to fetch lists"),
+            life: 3000,
+        });
+    }
+};
+
+const createList = async () => {
+    try {
+        const response = await fetch(arches.urls.controlled_list_add, {
+            method: "POST",
+            headers: {
+                "X-CSRFToken": Cookies.get("csrftoken"),
+            },
+        });
+        if (response.ok) {
+            const newItem = await response.json();
+            lists.value.unshift(newItem);
+        } else {
+            throw new Error();
+        }
+    } catch {
+        toast.add({
+            severity: ERROR,
+            summary: $gettext("List creation failed"),
+            life: 3000,
+        });
+    }
+};
+
+const deleteLists = async (selectedLists: ControlledList[]) => {
+    if (!selectedLists.length) {
+        return;
+    }
+    const promises = selectedLists.map((list) =>
+        fetch(arches.urls.controlled_list(list.id), {
+            method: "DELETE",
+            headers: {
+                "X-CSRFToken": Cookies.get("csrftoken"),
+            },
+        })
+    );
+
+    const shouldResetDisplay = (
+        selectedLists.some(l => l.id === displayedList.value?.id)
+    );
+
+    try {
+        const responses = await Promise.all(promises);
+        if (responses.some((resp) => resp.ok)) {
+            if (shouldResetDisplay) {
+                displayedList.value = null;
+            }
+        }
+        responses.forEach(async (response) => {
+            if (!response.ok) {
+                const body = await response.json();
+                toast.add({
+                    severity: ERROR,
+                    summary: $gettext("List deletion failed"),
+                    detail: body.message,
+                    life: 8000,
+                });
+            }
+        });
+    } catch {
+        toast.add({
+            severity: ERROR,
+            summary: $gettext("List deletion failed"),
+            life: 5000,
+        });
+    }
+    await fetchLists();
+};
+
+await fetchLists();
 </script>
 
 <template>
     <SearchAddDelete
         v-model="searchValue"
-        :create-action
-        :add-label
+        :create-action="createList"
         :del-action="
             () => {
-                delAction(selected);
+                deleteLists(selected);
                 selected.splice(0);
             }
         "
-        :del-label
-        :del-label-plural
         :number-to-delete="selected.length"
     />
     <div class="selection-header">
         <span
-            v-if="selectables.length"
+            v-if="lists.length"
             style="margin-left: 1rem"
         >
             <button
@@ -112,17 +184,17 @@ await props.fetchAction();
             </button>
         </span>
         <span
-            v-if="selectables.length"
+            v-if="lists.length"
             style="margin-right: 1rem"
         >
-            {{ selectables.length }}
-            {{ countLabel }}
+            {{ lists.length }}
+            {{ $ngettext("list", "lists", lists.length) }}
         </span>
     </div>
 
     <DataView
-        v-if="selectables.length"
-        :value="filteredSelectables"
+        v-if="lists.length"
+        :value="filteredLists"
     >
         <template #list="slotProps">
             <div
@@ -144,7 +216,7 @@ await props.fetchAction();
         </template>
         <template #empty>
             <div>
-                <span class="no-selections">{{ noSearchResultLabel }}</span>
+                <span class="no-selections">{{ NO_MATCHING_ITEMS }}</span>
             </div>
         </template>
     </DataView>
@@ -153,7 +225,7 @@ await props.fetchAction();
         v-else
         class="no-selections"
     >
-        <span>{{ noSelectionLabel }}</span>
+        <span>{{ NO_SELECTION_LABEL }}</span>
     </div>
 </template>
